move files to misc
