+++ AlphaTwirl/__init__.py
@@ -12,8 +12,8 @@ import Loop
 import ProgressBar
 import Aggregate
 import Concurrently
-from mkdir_p import mkdir_p
-from listToAlignedText import listToAlignedText
+from misc import mkdir_p
+from misc import listToAlignedText
 
 ##__________________________________________________________________||
 import logging
