+++ tests/unit/misc/test_listToAlignedText.py
